fix(audit): resolve appendix findings + NJ items across all modules
Works through the 110 unverified appendix findings (and the deferred
needs-judgment items NJ4 + random split_key backup) on top of the
confirmed-bug pass. Each behavioral fix carries a regression test that
fails pre-fix and passes post-fix; documented-intent and false-positive
findings are left as-is. Runtime/user-input validation that used `assert`
is converted to `raise` so it survives `python -O`.

core: _compatible_import.wraps copies each metadata attribute independently;
_deprecation drops dead code and uses the lazy replacement-module property;
_state init-hook metadata excludes internal bookkeeping, copy() gets an
independent hook manager, numel() raises on a poisoned state, StateTraceStack
exit verifies stack ownership; Hook id counter RMW is lock-guarded;
has_hooks validates type up front; environ.reset clears per-key locks and
tolerance honors the env precision dtype; mixin HashableDict hashes via
frozenset, JointTypes alias call raises; typing PyTree validates leaf
hashability, normalizes whitespace, and rejects malformed ellipsis structures.

graph: removed write-only context stacks; convert/flatten/walk assert->raise
with descriptive errors; dropped a redundant Everything predicate.

interop: AddScalar uses brainunit math so united Quantities work.

nn: softmin tuple-axis hint; soft_shrink/hard_shrink preserve integer dtype;
conv assert->raise + padding-parse guards; Threshold accepts Quantity;
removed dead extra_repr; MultiMetric reserves _metric_names; paddings accept
numpy/JAX integer scalars; poolings assert->raise, correct stride message,
channel_axis `is not None`, target_size validation; regularization sample_init
relu-guards the sqrt, validates group_size/order, and targets the requested
spectral norm; SoftplusT/NegSoftplusT demote via maybe_decimal; ScaledSigmoidT
rejects non-positive beta; EntropyReg/DirichletReg document the global-softmax
default (NJ4).

random: multinomial returns integer counts; zipf/geometric/triangular validate
parameters via jit_error_if; rayleigh(scale=None) no longer crashes; _numpy_keys
draws from the full uint32 range; the module-global DEFAULT key read-modify-write
is lock-guarded; split_key/self_assign_multi_keys back up the pre-split key so
restore_key actually rewinds (matches the documented contract).

transform: ifelse enforces the >=2-branch contract; _error_msg falls back to
str.format; fwd_grad guards integer dtypes and non-scalar outputs; grad NaN
check is inexact-aware and has_aux asserts->raise; IdentityMap iterates keys;
scan codegen uses collision-free skolem names; ir_visualize drops phantom pjit
edges; checkpointed_scan validates its two-tuple return; make_jaxpr guards
fn_to_check/None state_vals/keyword-only static_argnums, ignores weak_type-only
changes, and raises ValueError (not UnitMismatchError) on unit mismatch;
_remove_axis reports non-array leaves clearly; progress_bar assert->raise and
guards post-close updates; corrected unvmap/named_call docstrings; _util
assert->IndexError on argnums bounds.

util: split_total rejects bool; DotDict overrides __or__/__ior__; clear_buffer_memory
catches per-buffer; to_pure_dict keeps empty sub-dicts; StateJaxTracer is hashable;
filter WithTag rejects non-str tags and Any/All/Not hash defensively; documented
that a str filter matches by tag, not path.

+++ brainstate/random/_impl.py
@@ -92,9 +92,13 @@ def multinomial(
     if jnp.ndim(n) > 0:
         mask = _promote_shapes(jnp.arange(n_max) < jnp.expand_dims(n, -1), shape=shape + (n_max,))[0]
         mask = jnp.moveaxis(mask, -1, 0).astype(indices.dtype)
+        # ``excess`` must share ``indices``'s integer dtype: subtracting a
+        # float-typed ``excess`` (the previous ``jnp.zeros`` default) would promote
+        # the returned counts to floating point, violating the integer-counts
+        # contract for array-valued ``n``.
         excess = jnp.concatenate(
-            [jnp.expand_dims(n_max - n, -1),
-             jnp.zeros(u.math.shape(n) + (p.shape[-1] - 1,))],
+            [jnp.expand_dims((n_max - n).astype(indices.dtype), -1),
+             jnp.zeros(u.math.shape(n) + (p.shape[-1] - 1,), dtype=indices.dtype)],
             -1
         )
     else:
